Bug 1974334: Change SMILAnimationFunction::CompareTo to return int32_t instead of int8_t. r=longsonr
With this change to CompareTo, its return-type now matches that of
nsContentUtils::CompareTreePosition, which makes sense because we directly call
that function (and return the result) in CompareTo's final statement.

Previously, that call could trigger integer overflow (and the wrong comparison
outcome as a result).

Differential Revision: https://phabricator.services.mozilla.com/D255606

+++ dom/smil/SMILAnimationFunction.cpp
@@ -241,7 +241,7 @@ void SMILAnimationFunction::ComposeResult(const SMILAttr& aSMILAttr,
   }
 }
 
-int8_t SMILAnimationFunction::CompareTo(
+int32_t SMILAnimationFunction::CompareTo(
     const SMILAnimationFunction* aOther,
     nsContentUtils::NodeIndexCache& aCache) const {
   NS_ENSURE_TRUE(aOther, 0);
